# store-contract/src/main/resources/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Infinia Store Platform API
  version: 1.0.0
  summary: Catalog, publishing, review and delivery API for the Infinia / FengYu ecosystem.
  description: |
    Versioned REST API of the Infinia Store Platform (design §10).

    * Root path `/api/v1`; breaking changes ship as `/api/v2` — never via headers.
    * JSON is camelCase, timestamps are UTC RFC 3339, identifiers are UUIDv7/ULID.
    * Errors use `application/problem+json` with a stable `code`, `traceId` and
      localizable arguments.
    * Lists use cursor pagination; write endpoints accept `Idempotency-Key`;
      concurrent updates use `ETag/If-Match`.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.txt
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /releases/{releaseId}/install-manifest:
    get:
      operationId: getInstallManifest
      tags: [Delivery]
      summary: FengYu Native install manifest (aggregation plan §7.1) — signed artifact + install mode
      description: |
        Schema-versioned install contract for the host's SkillInstaller /
        McpInstaller. MCP installs always default to disabled with secrets
        kept local; compatibility surfaces are never a separate trust root.
      security:
        - {}
      parameters:
        - { name: releaseId, in: path, required: true, schema: { type: string } }
        - { name: client, in: query, schema: { type: string, default: fengyu } }
      responses:
        '200':
          description: Install manifest
          content:
            application/json:
              schema:
                type: object
                properties:
                  schemaVersion: { type: integer }
                  coordinate: { type: string }
                  type: { type: string }
                  sourceReleaseId: { type: string }
                  artifact:
                    type: object
                    properties:
                      url: { type: string, format: uri }
                      sha256: { type: string }
                      signature: { type: string }
                      keyId: { type: string }
                      size: { type: integer, format: int64 }
                  dependencies:
                    type: array
                    items:
                      type: object
                      properties:
                        coordinate: { type: string }
                        range: { type: string }
                        optional: { type: boolean }
                  permissions:
                    type: array
                    items: { $ref: '#/components/schemas/Permission' }
                  install:
                    type: object
                    properties:
                      mode: { type: string, enum: [SKILL_DIRECTORY, MCP_TEMPLATE, PLUGIN_PACKAGE] }
                      defaultEnabled: { type: boolean }
                      secretsPolicy: { type: string }
                  resolvedAt: { type: string, format: date-time }
        '409': { description: Release not installable }
  /compat/fengyu/mcp-catalog:
    get:
      operationId: fengyuMcpCatalog
      tags: [Compatibility]
      summary: Legacy-shape MCP catalog with direct template downloads
      security:
        - {}
      responses:
        '200':
          description: MCP catalog entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    name: { type: string }
                    description: { type: string }
                    version: { type: string }
                    author: { type: string }
                    downloadUrl: { type: string, description: Direct HMAC-ticketed template URL }
                    sha256: { type: string }
                    transport: { type: string }
                    official: { type: boolean }
  /compat/fengyu/codex/catalog:
    get:
      operationId: fengyuCodexCatalog
      tags: [Compatibility]
      summary: Codex-consumable skill catalog pointing at native manifests
      security:
        - {}
      responses:
        '200':
          description: Codex catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  schemaVersion: { type: integer }
                  generator: { type: string }
                  skills:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        skillId: { type: string }
                        version: { type: string }
                        listing: { type: string }
                        installManifest: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Permission:
      type: object
      properties:
        permissionId: { type: string }
        scope: { type: string, example: network }
        required: { type: boolean }
        reason: { type: string }
...